+++ lizard_workspace/management/commands/workspace_update.py
@@ -10,6 +10,7 @@ from lizard_workspace.tasks import (
     workspace_update_trackrecords,
     workspace_update_minimap,
     workspace_update_thememaps,
+    workspace_update_measure,
 )
 
 import optparse
@@ -38,6 +39,7 @@ class Command(BaseCommand):
         _option('trackrecords', 'Configure trackrecord layers'),
         _option('minimap', 'Configure minimap layers'),
         _option('thememaps', 'Configure thememaps layers'),
+        _option('thememapmeasure', 'Configure measure thememaps layers'),
         _option('all', 'All of the above options in one command'),
     )
 
@@ -53,3 +55,5 @@ class Command(BaseCommand):
             workspace_update_minimap(loglevel=10)
         if options.get('thememaps') or options.get('all'):
             workspace_update_thememaps(loglevel=10)
+        if options.get('thememapmeasure') or options.get('all'):
+            workspace_update_measure(loglevel=10)